REFACTOR: extract Arenas::Emission
any emission that is created in the context of an arena should inherit from Arenas::Emission
